+++ packages/block-editor/src/hooks/style.jsx
@@ -53,6 +53,7 @@ import { useBlockEditingMode } from '../components/block-editing-mode';
 import { useSettings } from '../components/use-settings';
 import { store as blockEditorStore } from '../store';
 import { globalStylesDataKey } from '../store/private-keys';
+import { isPlainObject } from '../utils/object';
 import { unlock } from '../lock-unlock';
 
 const { getResponsiveMediaQueries } = unlock( globalStylesEnginePrivateApis );
@@ -257,10 +258,6 @@ export function getStateStylesCSS( stateStyles, selector ) {
 	return [ importantCSS, textAlignCSS, fallbackCSS, backgroundResetCSS ]
 		.filter( Boolean )
 		.join( '\n' );
-}
-
-function isPlainObject( value ) {
-	return !! value && typeof value === 'object' && ! Array.isArray( value );
 }
 
 function mergeStyleObjects( target = {}, source = {} ) {