+++ packages/block-editor/src/hooks/utils.jsx
@@ -12,7 +12,11 @@ import {
 } from '../components/block-edit/context';
 import { useSettings } from '../components';
 import { useSettingsForBlockElement } from '../components/global-styles/hooks';
-import { getValueFromObjectPath, setImmutably } from '../utils/object';
+import {
+	getValueFromObjectPath,
+	isPlainObject,
+	setImmutably,
+} from '../utils/object';
 import { store as blockEditorStore } from '../store';
 import { unlock } from '../lock-unlock';
 
@@ -23,11 +27,7 @@ import { unlock } from '../lock-unlock';
  * @return {*} Object cleaned from falsy values
  */
 export const cleanEmptyObject = ( object ) => {
-	if (
-		object === null ||
-		typeof object !== 'object' ||
-		Array.isArray( object )
-	) {
+	if ( ! isPlainObject( object ) ) {
 		return object;
 	}